Kelvin Water Technologies Pvt. Ltd.
Kelvin Water Technologies Pvt. Ltd is one of the fastest and gradually growing water and wastewater treatment company in India. We work on multiple turnkeys project for water technologies and their treatment including waste management and wastewater treatment. We are committed to providing complete customer satisfaction. We manage our organization and water resources to meet evolving regulatory requirements, water supply needs and customer expectations in the future. We aim to be globally recognised as the first choice for water, wastewater, solid waste & organic waste services in the following communities.
For more information visit: https://www.kelvinindia.in/, Contact us on: +919812241001 or Email: info@kelvinindia.in